requirements.txt

python-docx
streamlit
PyYAML
openai>=1.0.0
